fhb_start

+++ src/main/scala/com/cpc/spark/ml/dnn/baseData/CollectIncTFData.scala
@@ -131,7 +131,7 @@ object MakeAdListV4Samples {
       rs =>
         ("min", rs._2)
     }).reduceByKey((x, y) => if (x < y) x else y).collectAsMap()
-    val min = min_map.getOrElse("min", -1)
+    val min = min_map.getOrElse("min", -1).toLong
     println("min idx of base instances file =" + min)
     val incremental_idx = min - 1
